// src/ReduxStore/favorities/interfaces.ts
export interface Movie {
  Actors:string;
  Awards: string;
  BoxOffice: string;
  Country: string;
  DVD: string;
  Director: string;
  Genre: string;
  Language: string;
  Metascore: string;
  Plot: string;
  Poster: string;
  Production: string;
  Rated: string;
  Ratings:any []
  Released: string;
  Response: string;
  Runtime: string;
  Title: string;
  Type: string;
  Website: string;
  Writer: string;
  Year: string;
  imdbID: string;
  imdbRating: string;
  imdbVotes: string;
}

export interface MovieState {
  favoriteMovies: Movie[];
}
export interface AddDeleteFavoriteMovieAction {
  type: 'ADD_DELETE_FAVORITE_MOVIE';
  movie: Movie;
}